requirements.txt
amqp==2.6.1
asgiref==3.6.0
billiard==3.6.4.0
celery==4.4.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-audiofield==0.10.1
django-celery==3.1.17
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
kombu==4.6.11
Markdown==3.4.3
prompt-toolkit==3.0.38
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.2
vine==1.3.0
wcwidth==0.2.6
